Check for None in search results entries
**Description**
This change is to address anomaly where an item in the results dict is None. This is unexpected response from the Spotify API - rather, it would be expected for the search request to return None for the entire query OR an empty list.

For now, a log message is added, and the loop continues such that this anomaly does not prevent `sync` from completing.

commit-id:53a3ac27